Fixed user object override.
When initializing the wp admin bar on a multisite, WP calls a function to switch blogs momentarily and then switch back (to detect user privs and such). So I had to have modify roles call a 2nd time on the action "switch_blog" to ensure the user object wasn't reset to default, but to the modified role.